+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Responsive, mobile support.
+- Mobile App icons and colors.
+- FavIcon
+
 ### Changed
 - Improved address menu to hide QR code on old/used addresses.
 - Improved address menu descriptions/tooltips.